requirements.txt
aerich==0.5.0
fastapi==0.63.0
pydantic==1.8.1
pypika-tortoise==0.1.0
pytz==2020.5
tortoise-orm==0.16.21
typing-extensions==3.7.4.3
asyncpg==0.24.0
